Rename batch into scheduler
