+++ packages/@react-facet/core/src/facet/createFacet.ts
@@ -1,6 +1,6 @@
 import { defaultEqualityCheck } from '../equalityChecks'
 import { Cleanup, EqualityCheck, Listener, WritableFacet, StartSubscription, Option, NO_VALUE } from '../types'
-import { batch } from '../batch'
+import { batch } from '../scheduler'
 
 export interface FacetOptions<V> {
   initialValue: Option<V>
